+++ src/main/java/testSetup/configreader/ConfigReader.java
@@ -1,6 +1,5 @@
 package testSetup.configreader;
 
-import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.util.Properties;
