+++ src/main/java/testSetup/deviceSetup/factory/ChromeDriverManager.java
@@ -3,13 +3,11 @@ package testSetup.deviceSetup.factory;
 import org.openqa.selenium.WebDriver;
 import testSetup.deviceSetup.ChromeDeviceSetup;
 
-public class ChromeDriverManager implements DriverManager{
+public class ChromeDriverManager extends DriverManager{
 
     @Override
     public WebDriver createDriver() throws Exception {
-//        WebDriverManager.chromedriver().setup();
-//        WebDriver driver = new ChromeDriver();
-//        driver.manage().window().maximize();
-        return ChromeDeviceSetup.driverBuilder();
+        driver = ChromeDeviceSetup.driverBuilder();
+        return driver;
     }
 }
