+++ src/main/java/testSetup/deviceSetup/factory/DriverManagerFactory.java
@@ -2,16 +2,14 @@ package testSetup.deviceSetup.factory;
 
 
 import testSetup.constants.TypesOfBrowsers;
-import testSetup.deviceSetup.factory.abstractFactory.DriverManagerAbstract;
-import testSetup.deviceSetup.factory.abstractFactory.FirefoxDriverManagerAbstract;
-import testSetup.deviceSetup.factory.abstractFactory.ChromeDriverManagerAbstract;
 
 public class DriverManagerFactory {
 
-    public static DriverManagerAbstract getManager(TypesOfBrowsers typesOfBrowsers){
+    public static DriverManager getManager(TypesOfBrowsers typesOfBrowsers){
         return switch (typesOfBrowsers) {
-            case CHROME -> new ChromeDriverManagerAbstract();
-            case FIREFOX -> new FirefoxDriverManagerAbstract();
+            case CHROME -> new ChromeDriverManager();
+            case FIREFOX -> new FirefoxDriverManager();
+            case EDGE -> new EdgeDriverManager();
             default -> throw new IllegalStateException("Unexpected value: " + typesOfBrowsers);
         };
     }
