+++ src/main/java/testSetup/deviceSetup/factory/EdgeDriverManager.java
@@ -1,16 +1,14 @@
 package testSetup.deviceSetup.factory;
 
-import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.edge.EdgeDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
 import testSetup.deviceSetup.EdgeDeviceSetup;
 
-public class EdgeDriverManager implements DriverManager{
+public class EdgeDriverManager extends DriverManager{
 
     @Override
     public WebDriver createDriver() throws Exception {
-
-        return EdgeDeviceSetup.driverBuilder();
+        driver = EdgeDeviceSetup.driverBuilder();
+        return driver;
     }
+
 }
