+++ src/main/java/testSetup/deviceSetup/factory/FirefoxDriverManager.java
@@ -1,14 +1,13 @@
 package testSetup.deviceSetup.factory;
 
-import io.github.bonigarcia.wdm.WebDriverManager;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.firefox.FirefoxDriver;
 import testSetup.deviceSetup.FirefoxDeviceSetup;
 
-public class FirefoxDriverManager implements DriverManager{
+public class FirefoxDriverManager extends DriverManager{
 
     @Override
     public WebDriver createDriver() throws Exception {
-        return FirefoxDeviceSetup.driverBuilder();
+        driver = FirefoxDeviceSetup.driverBuilder();
+        return driver;
     }
 }
