--- src/main/java/testSetup/deviceSetup/factory/abstractFactory/ChromeDriverManagerAbstract.java
@@ -1,11 +1,0 @@
-package testSetup.deviceSetup.factory.abstractFactory;
-
-import testSetup.deviceSetup.ChromeDeviceSetup;
-
-public class ChromeDriverManagerAbstract extends DriverManagerAbstract {
-
-    @Override
-    protected void startDriver() throws Exception {
-        driver  = ChromeDeviceSetup.driverBuilder();
-    }
-}
