--- src/main/java/testSetup/deviceSetup/factory/abstractFactory/DriverManagerFactoryAbstract.java
@@ -1,16 +1,0 @@
-package testSetup.deviceSetup.factory.abstractFactory;
-
-
-import testSetup.constants.TypesOfBrowsers;
-
-public class DriverManagerFactoryAbstract {
-
-    public static DriverManagerAbstract getManager(TypesOfBrowsers typesOfBrowsers){
-        return switch (typesOfBrowsers) {
-            case CHROME -> new ChromeDriverManagerAbstract();
-            case FIREFOX -> new FirefoxDriverManagerAbstract();
-            case EDGE -> new EdgeDriverManagerAbstract();
-            default -> throw new IllegalStateException("Unexpected value: " + typesOfBrowsers);
-        };
-    }
-}
