--- src/main/java/testSetup/deviceSetup/factory/abstractFactory/FirefoxDriverManagerAbstract.java
@@ -1,13 +1,0 @@
-package testSetup.deviceSetup.factory.abstractFactory;
-
-import io.github.bonigarcia.wdm.WebDriverManager;
-import org.openqa.selenium.firefox.FirefoxDriver;
-import testSetup.deviceSetup.FirefoxDeviceSetup;
-
-public class FirefoxDriverManagerAbstract extends DriverManagerAbstract {
-
-    @Override
-    protected void startDriver() throws Exception {
-        driver = FirefoxDeviceSetup.driverBuilder();
-    }
-}
